added skeleton code that should be tested
