+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utilities/CONFIG.java
@@ -18,6 +18,7 @@ public class CONFIG {
     public static String slidesMotor = "slidesMotor";
     public static String armMotor = "armMotor";
     public static String forearm = "clawR"; // clawRotator
+    public static String armServo = "armServo";
 
     // pull up (left is with the robot facing forward away from you)
     public static String leftPullUp = "cl";
